refactor: SearchPage 리팩터링
